feat 섭취 랭크 조회 기능 추가
